support current version of feedarser module, enabling python 3.9
commit df658c3 added some forward-compatibility with the then-current
development version of feedparser, which it assumed would be released
as feedparser 5.3.

As it turned out, the next feedparser release was 6.0.0, which
included another renaming of the `_resolveRelativeURIs` function
to `resolve_relative_uris`.

This commit adjusts the forward-compatibility code for that renaming,
corrects some comments referring to "feedparser 5.3" (which does not
exist, see above), and removes the `<6` constraint added to the
feedparser dependency by commit fa799c2e.

Being able to use feedparser version 6 is helpful because earlier
versions of that module fail to import with python >= 3.9:
https://github.com/kurtmckee/feedparser/issues/201