fix feedscanner crash with python >= 3.5
In python 3.3, html.parser's "non-strict" mode, which never raises
HTMLParseError, became the default. In python 3.5, the strict mode,
and the HTMLParseError class were both removed, causing the feed
scanning code to crash.

The test suite contains an example of HTML said to trigger a
HTMLParseError. With python 3.11 that HTML causes an AssertionError
inside html.parser, so this commit falls back to catching
AssertionError when HTMLParseError does not exist, so all python 3
versions should now behave correctly.

https://docs.python.org/3.3/library/html.parser.html#html.parser.HTMLParseError